기능: 지표 수동 집계 트리거 API 추가
POST /admin/metrics/aggregate - 지정 날짜(기본 전날) 집계 즉시 실행
?date=YYYY-MM-DD 파라미터로 특정 날짜 집계 가능
함께 CAST AS DOUBLE → CAST AS DOUBLE PRECISION PostgreSQL 호환 수정

+++ apps/admin-api/src/main/kotlin/com/publicplatform/ragops/adminapi/metrics/adapter/inbound/scheduler/MetricsAggregationScheduler.kt
@@ -36,7 +36,7 @@ class MetricsAggregationScheduler(
                 COUNT(q.id)                        AS total_questions,
                 SUM(CASE WHEN a.answer_status = 'fallback'  THEN 1 ELSE 0 END)  AS fallback_count,
                 SUM(CASE WHEN a.answer_status = 'no_answer' THEN 1 ELSE 0 END)  AS no_answer_count,
-                AVG(CAST(a.response_time_ms AS DOUBLE))                          AS avg_response_time_ms
+                AVG(CAST(a.response_time_ms AS DOUBLE PRECISION))               AS avg_response_time_ms
             FROM questions q
             LEFT JOIN answers a ON q.id = a.question_id
             WHERE CAST(q.created_at AS DATE) = :targetDate
